chatbot-project/backend/requirements.txt
fastapi
uvicorn
langchain-core
langchain-community
langchain-openai
python-dotenv
fastapi-cors
pydantic
python-dotenv
langchain
openai
sqlitedict
qdrant-client
pymongo
pinecone
sentence-transformers
pytz
